+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Iterable and Iterator implementations to `Scru128Generator` to make it work as
+  infinite iterator
+
 ## v2.2.0 - 2022-10-30
 
 ### Added